requirements.txt
ping3
pyyaml
